Merge stack #359 #363 #364

+++ engine/lib/phx/script/ffi_gen/InputDeviceType.lua
@@ -4,7 +4,7 @@ local Loader = {}
 
 function Loader.declareType()
     ffi.cdef [[
-        typedef uint8 InputDeviceType;
+        typedef uint16 InputDeviceType;
     ]]
 
     return 2, 'InputDeviceType'
